Split indico.modules.events.management.controllers
