+++ indico/modules/events/management/blueprint.py
@@ -17,13 +17,7 @@
 from __future__ import unicode_literals
 
 from indico.modules.events import event_management_object_url_prefixes
-from indico.modules.events.management.controllers import (RHEventSettings, RHEditEventData, RHEditEventDates,
-                                                          RHEditEventLocation, RHEditEventPersons,
-                                                          RHEditEventContactInfo, RHEditEventClassification,
-                                                          RHDeleteEvent, RHChangeEventType, RHLockEvent, RHUnlockEvent,
-                                                          RHShowNonInheriting, RHEventProtection, RHCloneEvent,
-                                                          RHMoveEvent, RHEventACL, RHEventACLMessage,
-                                                          RHPrintEventPoster, RHPosterPrintSettings, RHClonePreview)
+from indico.modules.events.management.controllers import cloning, posters, settings, actions, protection
 from indico.web.flask.util import make_compat_redirect_func
 from indico.web.flask.wrappers import IndicoBlueprint
 
@@ -32,26 +26,32 @@ _bp = IndicoBlueprint('event_management', __name__, template_folder='templates',
                       virtual_template_folder='events/management',
                       url_prefix='/event/<confId>/manage')
 
-_bp.add_url_rule('/', 'settings', RHEventSettings)
-_bp.add_url_rule('/settings/data', 'edit_data', RHEditEventData, methods=('GET', 'POST'))
-_bp.add_url_rule('/settings/dates', 'edit_dates', RHEditEventDates, methods=('GET', 'POST'))
-_bp.add_url_rule('/settings/location', 'edit_location', RHEditEventLocation, methods=('GET', 'POST'))
-_bp.add_url_rule('/settings/persons', 'edit_persons', RHEditEventPersons, methods=('GET', 'POST'))
-_bp.add_url_rule('/settings/contact-info', 'edit_contact_info', RHEditEventContactInfo, methods=('GET', 'POST'))
-_bp.add_url_rule('/settings/classification', 'edit_classification', RHEditEventClassification, methods=('GET', 'POST'))
-_bp.add_url_rule('/delete', 'delete', RHDeleteEvent, methods=('GET', 'POST'))
-_bp.add_url_rule('/change-type', 'change_type', RHChangeEventType, methods=('POST',))
-_bp.add_url_rule('/lock', 'lock', RHLockEvent, methods=('GET', 'POST'))
-_bp.add_url_rule('/unlock', 'unlock', RHUnlockEvent, methods=('POST',))
-_bp.add_url_rule('/protection', 'protection', RHEventProtection, methods=('GET', 'POST'))
-_bp.add_url_rule('/protection/acl', 'acl', RHEventACL)
-_bp.add_url_rule('/protection/acl-message', 'acl_message', RHEventACLMessage)
-_bp.add_url_rule('/move', 'move', RHMoveEvent, methods=('POST',))
-_bp.add_url_rule('/clone', 'clone', RHCloneEvent, methods=('GET', 'POST'))
-_bp.add_url_rule('/clone/preview', 'clone_preview', RHClonePreview, methods=('GET', 'POST'))
-
-_bp.add_url_rule('/print-poster/settings', 'poster_settings', RHPosterPrintSettings, methods=('GET', 'POST'))
-_bp.add_url_rule('/print-poster/<int:template_id>/<uuid>', 'print_poster', RHPrintEventPoster)
+# Settings
+_bp.add_url_rule('/', 'settings', settings.RHEventSettings)
+_bp.add_url_rule('/settings/data', 'edit_data', settings.RHEditEventData, methods=('GET', 'POST'))
+_bp.add_url_rule('/settings/dates', 'edit_dates', settings.RHEditEventDates, methods=('GET', 'POST'))
+_bp.add_url_rule('/settings/location', 'edit_location', settings.RHEditEventLocation, methods=('GET', 'POST'))
+_bp.add_url_rule('/settings/persons', 'edit_persons', settings.RHEditEventPersons, methods=('GET', 'POST'))
+_bp.add_url_rule('/settings/contact-info', 'edit_contact_info', settings.RHEditEventContactInfo,
+                 methods=('GET', 'POST'))
+_bp.add_url_rule('/settings/classification', 'edit_classification', settings.RHEditEventClassification,
+                 methods=('GET', 'POST'))
+# Actions
+_bp.add_url_rule('/delete', 'delete', actions.RHDeleteEvent, methods=('GET', 'POST'))
+_bp.add_url_rule('/change-type', 'change_type', actions.RHChangeEventType, methods=('POST',))
+_bp.add_url_rule('/lock', 'lock', actions.RHLockEvent, methods=('GET', 'POST'))
+_bp.add_url_rule('/unlock', 'unlock', actions.RHUnlockEvent, methods=('POST',))
+_bp.add_url_rule('/move', 'move', actions.RHMoveEvent, methods=('POST',))
+# Protection
+_bp.add_url_rule('/protection', 'protection', protection.RHEventProtection, methods=('GET', 'POST'))
+_bp.add_url_rule('/protection/acl', 'acl', protection.RHEventACL)
+_bp.add_url_rule('/protection/acl-message', 'acl_message', protection.RHEventACLMessage)
+# Cloning
+_bp.add_url_rule('/clone', 'clone', cloning.RHCloneEvent, methods=('GET', 'POST'))
+_bp.add_url_rule('/clone/preview', 'clone_preview', cloning.RHClonePreview, methods=('GET', 'POST'))
+# Posters
+_bp.add_url_rule('/print-poster/settings', 'poster_settings', posters.RHPosterPrintSettings, methods=('GET', 'POST'))
+_bp.add_url_rule('/print-poster/<int:template_id>/<uuid>', 'print_poster', posters.RHPrintEventPoster)
 
 
 for object_type, prefixes in event_management_object_url_prefixes.iteritems():
@@ -59,7 +59,7 @@ for object_type, prefixes in event_management_object_url_prefixes.iteritems():
         continue
     for prefix in prefixes:
         prefix = '!/event/<confId>' + prefix
-        _bp.add_url_rule(prefix + '/show-non-inheriting', 'show_non_inheriting', RHShowNonInheriting,
+        _bp.add_url_rule(prefix + '/show-non-inheriting', 'show_non_inheriting', protection.RHShowNonInheriting,
                          defaults={'object_type': object_type})
