+++ indico/modules/events/management/controllers/protection.py
@@ -1,0 +1,100 @@
+# This file is part of Indico.
+# Copyright (C) 2002 - 2017 European Organization for Nuclear Research (CERN).
+#
+# Indico is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License as
+# published by the Free Software Foundation; either version 3 of the
+# License, or (at your option) any later version.
+#
+# Indico is distributed in the hope that it will be useful, but
+# WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with Indico; if not, see <http://www.gnu.org/licenses/>.
+
+from __future__ import unicode_literals
+
+from flask import flash, redirect, request
+from werkzeug.exceptions import NotFound
+
+from indico.core.db.sqlalchemy.protection import render_acl, ProtectionMode
+from indico.modules.events.management.controllers.base import RHManageEventBase
+from indico.modules.events.management.forms import (EventProtectionForm)
+from indico.modules.events.management.views import (WPEventProtection)
+from indico.modules.events.operations import (update_event_protection)
+from indico.modules.events.sessions import session_settings, COORDINATOR_PRIV_SETTINGS
+from indico.modules.events.sessions.operations import update_session_coordinator_privs
+from indico.modules.events.util import get_object_from_args, update_object_principals
+from indico.util.i18n import _
+from indico.web.flask.util import url_for
+from indico.web.forms.base import FormDefaults
+from indico.web.util import jsonify_template
+
+
+class RHShowNonInheriting(RHManageEventBase):
+    """Show a list of non-inheriting child objects"""
+
+    def _checkParams(self, params):
+        RHManageEventBase._checkParams(self, params)
+        self.obj = get_object_from_args()[2]
+        if self.obj is None:
+            raise NotFound
+
+    def _process(self):
+        objects = self.obj.get_non_inheriting_objects()
+        return jsonify_template('events/management/non_inheriting_objects.html', objects=objects)
+
+
+class RHEventACL(RHManageEventBase):
+    """Display the inherited ACL of the event"""
+
+    def _process(self):
+        return render_acl(self.event_new)
+
+
+class RHEventACLMessage(RHManageEventBase):
+    """Render the inheriting ACL message"""
+
+    def _process(self):
+        mode = ProtectionMode[request.args['mode']]
+        return jsonify_template('forms/protection_field_acl_message.html', object=self.event_new, mode=mode,
+                                endpoint='event_management.acl')
+
+
+class RHEventProtection(RHManageEventBase):
+    """Show event protection"""
+
+    def _process(self):
+        form = EventProtectionForm(obj=FormDefaults(**self._get_defaults()), event=self.event_new)
+        if form.validate_on_submit():
+            update_event_protection(self.event_new, {'protection_mode': form.protection_mode.data,
+                                                     'own_no_access_contact': form.own_no_access_contact.data,
+                                                     'access_key': form.access_key.data,
+                                                     'visibility': form.visibility.data})
+            update_object_principals(self.event_new, form.acl.data, read_access=True)
+            update_object_principals(self.event_new, form.managers.data, full_access=True)
+            update_object_principals(self.event_new, form.submitters.data, role='submit')
+            self._update_session_coordinator_privs(form)
+            flash(_('Protection settings have been updated'), 'success')
+            return redirect(url_for('.protection', self.event_new))
+        return WPEventProtection.render_template('event_protection.html', self.event_new, 'protection', form=form)
+
+    def _get_defaults(self):
+        acl = {p.principal for p in self.event_new.acl_entries if p.read_access}
+        submitters = {p.principal for p in self.event_new.acl_entries if p.has_management_role('submit', explicit=True)}
+        managers = {p.principal for p in self.event_new.acl_entries if p.full_access}
+        registration_managers = {p.principal for p in self.event_new.acl_entries
+                                 if p.has_management_role('registration', explicit=True)}
+        event_session_settings = session_settings.get_all(self.event_new)
+        coordinator_privs = {name: event_session_settings[val] for name, val in COORDINATOR_PRIV_SETTINGS.iteritems()
+                             if event_session_settings.get(val)}
+        return dict({'protection_mode': self.event_new.protection_mode, 'acl': acl, 'managers': managers,
+                     'registration_managers': registration_managers, 'submitters': submitters,
+                     'access_key': self.event_new.access_key, 'visibility': self.event_new.visibility,
+                     'own_no_access_contact': self.event_new.own_no_access_contact}, **coordinator_privs)
+
+    def _update_session_coordinator_privs(self, form):
+        data = {field: getattr(form, field).data for field in form.priv_fields}
+        update_session_coordinator_privs(self.event_new, data)
